[REF] Refactored project structure, rm deprc API
- Completely refactored project structure, especially
created more header files to boost code-readability

- Removed deprecated and unused functions, removed comments
