+++ src/binary_api/shifts/s21_binary_shifts.h
@@ -1,0 +1,9 @@
+#ifndef S21_BINARY_SHIFTS_H_
+#define S21_BINARY_SHIFTS_H_
+
+#include "../s21_structures.h"
+
+void shiftr(s21_decimal *a);
+int shiftl(s21_decimal *a);
+
+#endif  // S21_BINARY_SHIFTS_H_
