+++ src/math_ops/s21_mod.c
@@ -1,6 +1,5 @@
 #include "../s21_decimal.h"
 
-
 void handle_exponent_mod(s21_decimal value_1, s21_decimal value_2,
                          s21_decimal *result, int *code);
