+++ src/s21_decimal.h
@@ -8,31 +8,9 @@
 #include <string.h>
 
 #include "binary_api/s21_core.h"
+#include "comparison/s21_comparison.h"
+#include "conversion/s21_conversion.h"
+#include "math_ops/s21_dec_math.h"
 #include "rounding/s21_rounding.h"
-
-/* Smart artithmetic API */
-int s21_add(s21_decimal value_1, s21_decimal value_2, s21_decimal *result);
-int s21_sub(s21_decimal value_1, s21_decimal value_2, s21_decimal *result);
-int s21_mul(s21_decimal value_1, s21_decimal value_2, s21_decimal *result);
-int s21_div(s21_decimal value_1, s21_decimal value_2, s21_decimal *result);
-int s21_mod(s21_decimal value_1, s21_decimal value_2, s21_decimal *result);
-
-/* Basic arithmetic API */
-s21_decimal s21_integer_mod(s21_decimal dividend, s21_decimal divisor);
-s21_decimal s21_integer_div(s21_decimal dividend, s21_decimal divisor,
-                            s21_decimal *result);
-
-int s21_is_less(s21_decimal a, s21_decimal b);
-int s21_is_less_or_equal(s21_decimal a, s21_decimal b);
-int s21_is_greater(s21_decimal a, s21_decimal b);
-int s21_is_greater_or_equal(s21_decimal a, s21_decimal b);
-int s21_is_equal(s21_decimal a, s21_decimal b);
-int s21_is_not_equal(s21_decimal a, s21_decimal b);
-
-int s21_from_int_to_decimal(int src, s21_decimal *dst);
-int s21_from_float_to_decimal(float src, s21_decimal *dst);
-int s21_from_decimal_to_int(s21_decimal src, int *dst);
-int s21_from_decimal_to_float(s21_decimal src, float *dst);
-int s21_negate(s21_decimal value, s21_decimal *result);
 
 #endif  //  SRC_S21_DECIMAL_H_
