--- src/tests/binary_tests/test_s21_binary_api.c
@@ -1,24 +1,0 @@
-/**
- * @file test_s21_binary_api.c
- * @author Mikhail Kuznetsov (mikhail.n.kuznetsov@gmail.com)
- * @brief This file exists to boost test coverage
- * of certain binary API functions.
- * @version 0.1
- * @date 2022-06-10
- *
- * @copyright Copyright (c) 2022
- */
-
-#include "../s21_decimal_test.h"
-
-// TODO(tartica): implement this file to boost gcov % test coverage
-
-// Suite *suite_convert_decimal_to_gmp(void) {
-//     Suite *s = suite_create(PRETTY_PRINT("s21_convert_decimal_to_gmp"));
-//     TCase *tc = tcase_create("s21_convert_decimal_to_gmp");
-
-//     tcase_add_loop_test(tc, convert_test, 0, 100);
-
-//     suite_add_tcase(s, tc);
-//     return s;
-// }
