--- src/tests/helpers/gmp/gmp_arithmetic_wrappers.c
@@ -1,5 +1,0 @@
-#include "../../s21_decimal_test.h"
-
-// void gmp_sum_int(mpz_t *res, int64_t compound_val) {
-//     mpz_add_ui(*res, *res, compound_val);
-// }
