+++ src/tests/helpers/printers.c
@@ -10,25 +10,6 @@ void print_mpz_decimal(mpz_t mpz_val) {
     puts("");
 }
 
-// void print_bits(s21_decimal d) {
-//     for (int i = 0; i < 4; i++) {
-//         if (i == 2)
-//             printf(" ");
-//         int x = 3;
-//         for (int j = 0; j < 32; j++) {
-//             if (IS_SET(d.bits[i], j))
-//                 printf("%s", HRED);
-//             printf("%u", IS_SET(d.bits[i], j));
-//             if (j == x && !(i == 2 && j == 31)) {
-//                 x += 4;
-//                 // printf("");
-//             }
-//             printf("%s", RESET);
-//         }
-//     }
-//     printf("\n");
-// }
-
 void print_bits_no_exp(s21_decimal d) {
     int no_set = 0;
 
@@ -72,35 +53,3 @@ void print_bits_r(s21_decimal d) {
     }
     printf("\n");
 }
-
-// void print_bits_r_set(s21_decimal d, int set_n) {
-//     int x = 28;
-//     for (int j = 15; j >= 0; --j) {
-//         if (IS_SET(d.bits[set_n], j))
-//             printf("%s", HRED);
-//         printf("%u", IS_SET(d.bits[set_n], j));
-//         if (j == x && !(set_n == 0 && j == 0)) {
-//             x -= 4;
-//             // printf("");
-//         }
-//         printf("%s", RESET);
-//     }
-
-//     printf("\n");
-// }
-
-// void print_bits_set(s21_decimal d, int set_n) {
-//     int x = 28;
-//     for (int j = 0; j < 16; ++j) {
-//         if (IS_SET(d.bits[set_n], j))
-//             printf("%s", HRED);
-//         printf("%u", IS_SET(d.bits[set_n], j));
-//         if (j == x && !(set_n == 0 && j == 0)) {
-//             x -= 4;
-//             // printf("");
-//         }
-//         printf("%s", RESET);
-//     }
-
-//     printf("\n");
-// }
