+++ src/tests/test_runner.c
@@ -33,28 +33,28 @@ void run_testcase(Suite *testcase) {
 
 void run_tests(void) {
     Suite *list_cases[] = {
-        //    suite_s21_add(),  // ⚠️
-        //    suite_s21_mul(),                    // ⚠️
+           suite_s21_add(),  // ⚠️
+           suite_s21_mul(),                    // ⚠️
 
-        //    suite_s21_round(),                  // ⚠️
+           suite_s21_round(),                  // ⚠️
            suite_s21_mod(),                    // ⚠️
            suite_s21_div(),                    // ⚠️
-        //    suite_s21_from_float_to_decimal(),  // ⚠️
-        //    suite_s21_floor(),                  // ⚠️
-        //    suite_s21_from_decimal_to_float(),  // ⚠️
+           suite_s21_from_float_to_decimal(),  // ⚠️
+           suite_s21_floor(),                  // ⚠️
+           suite_s21_from_decimal_to_float(),  // ⚠️
 
-        //    suite_s21_sub(),                    // ✅
-        //    suite_s21_is_less(),                // ✅
-        //    suite_s21_truncate(),               // ✅
-        //    suite_s21_from_int_to_decimal(),    // ✅
-        //    suite_s21_is_less_or_equal(),       // ✅
-        //    suite_s21_is_greater(),             // ✅
-        //    suite_s21_is_greater_or_equal(),    // ✅
-        //    suite_s21_is_or_not_equal(),        // ✅
-        //    suite_convert_gmp_to_decimal(),     // ✅
-        //    suite_convert_decimal_to_gmp(),     // ✅
-        //    suite_s21_from_decimal_to_int(),    // ✅
-        //    suite_s21_negate(),                 // ✅
+           suite_s21_sub(),                    // ✅
+           suite_s21_is_less(),                // ✅
+           suite_s21_truncate(),               // ✅
+           suite_s21_from_int_to_decimal(),    // ✅
+           suite_s21_is_less_or_equal(),       // ✅
+           suite_s21_is_greater(),             // ✅
+           suite_s21_is_greater_or_equal(),    // ✅
+           suite_s21_is_or_not_equal(),        // ✅
+           suite_convert_gmp_to_decimal(),     // ✅
+           suite_convert_decimal_to_gmp(),     // ✅
+           suite_s21_from_decimal_to_int(),    // ✅
+           suite_s21_negate(),                 // ✅
 
         NULL};
